moved profile to urls and view to their correct place
