+++ imagersite/imager_profile/templates/imager_profile/profile.html
@@ -4,7 +4,7 @@
 <h2>Welcome {{user.username}}!</h2>
 <ul>
     <li>Name: {{user.first_name}} {{user.last_name}}</li>
-    <li>Email: <a href="mailto:{{user.email}}" "email me">{{user.email}}</a></li>
+    <li>Email: <a href="mailto:{{user.email}}">{{user.email}}</a></li>
     <li>Bio: {{user.profile.bio}}</li>
     <li>Website: <a href="{{user.profile.personal_website}}">{{user.profile.personal_website}}</a></li>
     <li>For Hire: {{user.profile.for_hire}}</li>
